% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM11.R
\name{admm_algorithm}
\alias{admm_algorithm}
\title{ADMM Algorithm for Constrained Optimization}
\usage{
admm_algorithm(A, B, C, f, g, rho = 1, max_iter = 100, tol = 1e-06)
}
\arguments{
\item{A}{A numeric matrix.}

\item{B}{A numeric matrix.}

\item{C}{A numeric matrix.}

\item{f}{A equation for X.}

\item{g}{A equation for X.}

\item{rho}{A numeric scalar representing the penalty parameter for ADMM (default is 1).}

\item{max_iter}{An integer representing the maximum number of iterations (default is 100).}

\item{tol}{A numeric scalar representing the tolerance for convergence (default is 1e-6).}
}
\value{
A list containing:
  \item{x}{The optimized \eqn{x} values.}
  \item{z}{The optimized \eqn{z} values.}
  \item{lambda}{The final Lagrange multipliers.}
}
\description{
Find the minimum value of f(x)+g(z) under the condition Ax+Bz=C
}
\examples{
\dontrun{
A <- matrix(c(1, 2, 3, 4), 2, 2)
B <- diag(2)
C <- c(1, 1)
f <- function(x) sum(x^2)
g <- function(z) sum(abs(z))
result <- admm_algorithm(A, B, C, f, g)
print(result$x)
}
}
